+++ Robust.Client/GameObjects/ClientEntityManager.cs
@@ -29,6 +29,9 @@ namespace Robust.Client.GameObjects
         internal event Action? AfterStartup;
         internal event Action? AfterShutdown;
 
+        private readonly Queue<EntityUid> _queuedPredictedDeletions = new();
+        private readonly HashSet<EntityUid> _queuedPredictedDeletionsSet = new();
+
         public override void Initialize()
         {
             SetupNetworking();
@@ -213,6 +216,34 @@ namespace Robust.Client.GameObjects
                 }
             }
 
+            using (histogram?.WithLabels("PredictedQueueDel").NewTimer())
+            {
+                while (_queuedPredictedDeletions.TryDequeue(out var uid))
+                {
+                    if (!MetaQuery.TryGetComponentInternal(uid, out var meta))
+                        continue;
+
+                    if (meta.EntityLifeStage >= EntityLifeStage.Terminating)
+                        continue;
+
+                    var xform = TransformQuery.GetComponentInternal(uid);
+                    if (meta.NetEntity.IsClientSide())
+                    {
+                        DeleteEntity(uid, meta, xform);
+                    }
+                    else
+                    {
+                        _xforms.DetachEntity(uid, xform, meta, null);
+                        // base call bypasses IGameTiming.InPrediction check
+                        // This is pretty janky and there should be a way for the client to dirty an entity outside of prediction
+                        // TODO PREDICTION
+                        base.Dirty(uid, xform, meta);
+                    }
+                }
+
+                _queuedPredictedDeletionsSet.Clear();
+            }
+
             base.TickUpdate(frameTime, noPredictions, histogram);
         }
 
@@ -317,18 +348,23 @@ namespace Robust.Client.GameObjects
             }
         }
 
-        /// <inheritdoc />
-        public override void PredictedQueueDeleteEntity(Entity<MetaDataComponent?, TransformComponent?> ent)
-        {
-            if (IsQueuedForDeletion(ent.Owner)
-                || !MetaQuery.Resolve(ent.Owner, ref ent.Comp1)
-                || ent.Comp1.EntityLifeStage >= EntityLifeStage.Terminating
-                || !TransformQuery.Resolve(ent.Owner, ref ent.Comp2))
-            {
-                return;
-            }
+        public override bool IsQueuedForDeletion(EntityUid uid)
+            => QueuedDeletionsSet.Contains(uid) || _queuedPredictedDeletions.Contains(uid);
 
-            if (ent.Comp1.NetEntity.IsClientSide())
+        /// <inheritdoc />
+        public override void PredictedQueueDeleteEntity(Entity<MetaDataComponent?> ent)
+        {
+            // Some UIs get disposed after entity-manager has shut down and already deleted all entities.
+            if (!Started)
+                return;
+
+            if (IsQueuedForDeletion(ent.Owner))
+                return;
+
+            if (!MetaQuery.Resolve(ent.Owner, ref ent.Comp, false))
+                return;
+
+            if (ent.Comp.NetEntity.IsClientSide())
             {
                 // client-side QueueDeleteEntity re-fetches MetadataComp and checks IsClientSide().
                 // base call to skip that.
@@ -337,7 +373,10 @@ namespace Robust.Client.GameObjects
             }
             else
             {
-                _xforms.DetachEntity(ent.Owner, ent.Comp2);
+                if (!_queuedPredictedDeletionsSet.Add(ent.Owner))
+                    return;
+
+                _queuedPredictedDeletions.Enqueue(ent.Owner);
             }
         }
     }