+++ Robust.Shared/GameObjects/EntityManager.cs
@@ -694,7 +694,7 @@ namespace Robust.Shared.GameObjects
             EntityQueueDeleted?.Invoke(uid.Value);
         }
 
-        public bool IsQueuedForDeletion(EntityUid uid) => QueuedDeletionsSet.Contains(uid);
+        public virtual bool IsQueuedForDeletion(EntityUid uid) => QueuedDeletionsSet.Contains(uid);
 
         /// <inheritdoc />
         public virtual void PredictedDeleteEntity(Entity<MetaDataComponent?, TransformComponent?> ent)
@@ -712,18 +712,44 @@ namespace Robust.Shared.GameObjects
         }
 
         /// <inheritdoc />
-        public virtual void PredictedQueueDeleteEntity(Entity<MetaDataComponent?, TransformComponent?> ent)
+        public virtual void PredictedQueueDeleteEntity(Entity<MetaDataComponent?> ent)
         {
-            QueueDeleteEntity(ent.Owner);
+            QueueDeleteEntity(ent);
         }
 
         /// <inheritdoc />
-        public virtual void PredictedQueueDeleteEntity(Entity<MetaDataComponent?, TransformComponent?>? ent)
+        public void PredictedQueueDeleteEntity(Entity<MetaDataComponent?>? ent)
         {
-            if (ent == null)
-                return;
+            if (ent != null)
+                PredictedQueueDeleteEntity(ent.Value);
+        }
 
-            PredictedQueueDeleteEntity(ent.Value);
+        /// <inheritdoc />
+        [Obsolete("use variant without TransformComponent")]
+        public void PredictedQueueDeleteEntity(Entity<MetaDataComponent?, TransformComponent?> ent)
+        {
+            PredictedQueueDeleteEntity(new Entity<MetaDataComponent?>(ent.Owner, ent.Comp1));
+        }
+
+        /// <inheritdoc />
+        [Obsolete("use variant without TransformComponent")]
+        public void PredictedQueueDeleteEntity(Entity<MetaDataComponent?, TransformComponent?>? ent)
+        {
+            if (ent != null)
+                PredictedQueueDeleteEntity(new Entity<MetaDataComponent?>(ent.Value.Owner, ent.Value.Comp1));
+        }
+
+        /// <inheritdoc />
+        public void PredictedQueueDeleteEntity(EntityUid uid)
+        {
+            PredictedQueueDeleteEntity(new Entity<MetaDataComponent?>(uid, null));
+        }
+
+        /// <inheritdoc />
+        public void PredictedQueueDeleteEntity(EntityUid? uid)
+        {
+            if (uid != null)
+                PredictedQueueDeleteEntity(new Entity<MetaDataComponent?>(uid.Value, null));
         }
 
         public bool EntityExists(EntityUid uid)