+++ Robust.Shared/GameObjects/IEntityManager.cs
@@ -167,20 +167,31 @@ namespace Robust.Shared.GameObjects
         /// </summary>
         void PredictedDeleteEntity(Entity<MetaDataComponent?, TransformComponent?> ent);
 
-        /// <summary>
-        /// <see cref="PredictedDeleteEntity(Robust.Shared.GameObjects.Entity{Robust.Shared.GameObjects.MetaDataComponent?,Robust.Shared.GameObjects.TransformComponent?})"/>
-        /// </summary>
+        /// <inheritdoc cref="PredictedDeleteEntity(Entity{MetaDataComponent?,TransformComponent?})"/>
         void PredictedDeleteEntity(Entity<MetaDataComponent?, TransformComponent?>? ent);
 
         /// <summary>
-        /// Tries to predict entity deletion. On the server it runs the normal code path and on the client the entity is detached.
+        /// Variant of <see cref="PredictedDeleteEntity(Entity{MetaDataComponent?,TransformComponent?})"/> that defers deletion until the end of the tick.
         /// </summary>
+        void PredictedQueueDeleteEntity(Entity<MetaDataComponent?> ent);
+
+        /// <inheritdoc cref="PredictedQueueDeleteEntity(Entity{MetaDataComponent?})"/>
+        void PredictedQueueDeleteEntity(Entity<MetaDataComponent?>? ent);
+
+        /// <inheritdoc cref="PredictedQueueDeleteEntity(Entity{MetaDataComponent?})"/>
+        [Obsolete("use variant without TransformComponent")]
         void PredictedQueueDeleteEntity(Entity<MetaDataComponent?, TransformComponent?> ent);
 
-        /// <summary>
-        /// <see cref="PredictedQueueDeleteEntity(Robust.Shared.GameObjects.Entity{Robust.Shared.GameObjects.MetaDataComponent?,Robust.Shared.GameObjects.TransformComponent?})"/>
-        /// </summary>
+        /// <inheritdoc cref="PredictedQueueDeleteEntity(Entity{MetaDataComponent?})"/>
+        [Obsolete("use variant without TransformComponent")]
         void PredictedQueueDeleteEntity(Entity<MetaDataComponent?, TransformComponent?>? ent);
+
+        // The following 2 methods exist to handle ambiguous implicit casts.
+        // They should probably be removed whenever the transform variants are removed
+        /// <inheritdoc cref="PredictedQueueDeleteEntity(Entity{MetaDataComponent?})"/>
+        void PredictedQueueDeleteEntity(EntityUid ent);
+        /// <inheritdoc cref="PredictedQueueDeleteEntity(Entity{MetaDataComponent?})"/>
+        void PredictedQueueDeleteEntity(EntityUid? ent);
 
         /// <summary>
         /// Shuts-down and removes the entity with the given <see cref="Robust.Shared.GameObjects.EntityUid"/>. This is also broadcast to all clients.