+++ RELEASE-NOTES.md
@@ -39,11 +39,35 @@ END TEMPLATE-->
 
 ### New features
 
+*None yet*
+
+### Bugfixes
+
+* Fixed yaml hot reloading throwing invalid path exceptions.
+
+### Other
+
+*None yet*
+
+### Internal
+
+*None yet*
+
+
+## 267.2.1
+
+
+## 267.2.0
+
+### New features
+
 * Sprites and Sprite layers have a new `Loop` data field that can be set to false to automatically pause animations once they have finished.
 
 ### Bugfixes
 
 * Fixed `CollectionExtensions.TryGetValue` throwing an exception when given a negative list index.
+* Fixed `EntityManager.PredictedQueueDeleteEntity()` not deferring changes for networked entities until the end of the tick. 
+* Fixed `EntityManager.IsQueuedForDeletion` not returning true foe entities getting deleted via `PredictedQueueDeleteEntity()`
 
 ### Other